// src/app/shared/interfaces/appSettings.interface.ts
export interface AppSettings {
  googleAnalyticsJqueryUrl: string;
  googleAnalyticsJsUrl: string;
  googleAnalyticsProfileUrl: string;
  ocelotApiGatewayBaseUrl: string;
  resourcesApiBaseUrl: string,
  projectsPageSize: number;
  resourcesPerPage: number;
  scrollDistance: number,
  staffingAuthenticationApiBaseUrl: string;
  clearLocalStorage: string;
  siteUnderMaintainance: string;
  launchDarklyClientId;
}